update : quiz/problem 삭제 로직 s3 추가

+++ src/main/java/com/dsm/newtrash/back/springboot/domain/quiz/service/QuizService.java
@@ -3,6 +3,7 @@ package com.dsm.newtrash.back.springboot.domain.quiz.service;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.dsm.newtrash.back.springboot.domain.problem.exception.ProblemNotWriteUserException;
 import com.dsm.newtrash.back.springboot.domain.problem.presentation.dto.request.ProblemRequest;
 import com.dsm.newtrash.back.springboot.domain.problem.service.QuizProblemService;
 import com.dsm.newtrash.back.springboot.domain.quiz.domain.Quiz;
@@ -11,6 +12,7 @@ import com.dsm.newtrash.back.springboot.domain.quiz.domain.type.CategoryType;
 import com.dsm.newtrash.back.springboot.domain.quiz.exception.QuizNotFoundException;
 import com.dsm.newtrash.back.springboot.domain.quiz.presentation.dto.request.QuizRequest;
 import com.dsm.newtrash.back.springboot.domain.user.service.util.UserUtil;
+import com.dsm.newtrash.back.springboot.global.aws.S3Util;
 
 import lombok.RequiredArgsConstructor;
 
@@ -20,6 +22,7 @@ public class QuizService {
 
 	private final QuizRepository quizRepository;
 	private final UserUtil userUtil;
+	private final S3Util s3Util;
 	private final QuizProblemService quizProblemService;
 
 	@Transactional(rollbackFor = Exception.class)
@@ -41,8 +44,12 @@ public class QuizService {
 	@Transactional
 	public void deleteQuiz(Long quizId) {
 		Quiz quiz = quizRepository.findById(quizId).orElseThrow(() -> QuizNotFoundException.EXCEPTION);
-		quizRepository.delete(quiz);
+
+		if(!quiz.getUser().equals(userUtil.getUser())) throw ProblemNotWriteUserException.EXCEPTION;
+
+		s3Util.delete(quiz.getPath());
 		quizProblemService.deleteAllProblem(quizId);
+		quizRepository.delete(quiz);
 	}
 
 }
